+++ Source/Meadow.Foundation.Peripherals/Displays.TftSpi/Driver/Drivers/ST7735.cs
@@ -295,14 +295,14 @@ namespace Meadow.Foundation.Displays
                 0x1B, 0x1A, 0x24, 0x2B, 0x06, 0x06, 0x02, 0x0F,
             });
 
-            SendCommand((byte)LcdCommand.CASET);
+            SendCommand(LcdCommand.CASET);
             SendData(new byte[]
             {
                 0x00, 0x02,             //     XSTART = 2
                 0x00, 0x81,             //     XEND = 129
             });
 
-            SendCommand((byte)LcdCommand.RASET);
+            SendCommand(LcdCommand.RASET);
             SendData(new byte[]
             {
                 0x00, 0x02,             //     XSTART = 1
@@ -385,21 +385,21 @@ namespace Meadow.Foundation.Displays
             x1 += xOffset;
             y1 += yOffset;
 
-            SendCommand((byte)LcdCommand.CASET);  // column addr set
+            SendCommand(LcdCommand.CASET);  // column addr set
             dataCommandPort.State = Data;
             Write((byte)(x0 >> 8));
             Write((byte)(x0 & 0xff));   // XSTART 
             Write((byte)(x1 >> 8));
             Write((byte)(x1 & 0xff));   // XEND
 
-            SendCommand((byte)LcdCommand.RASET);  // row addr set
+            SendCommand(LcdCommand.RASET);  // row addr set
             dataCommandPort.State = Data;
             Write((byte)(y0 >> 8));
             Write((byte)(y0 & 0xff));   // YSTART 
             Write((byte)(y1 >> 8));
             Write((byte)(y1 & 0xff));   // YEND
 
-            SendCommand((byte)LcdCommand.RAMWR);  // write to RAM
+            SendCommand(LcdCommand.RAMWR);  // write to RAM
         }
     }
 }
